Calculate Segment Length Based on Specific Time Period

LP1 allows you calculate where to cut to represent a specific time period. It uses the discrete step by step von Bertanlanffy equation solved for L(p-1)

LP1(L, L.asym, Time, k)

Arguments

L
numeric; length of the sample

L.asym
numeric; asymptotic length

Time
numeric; time period in days to be represented by the cut segment

k
numeric; growth coefficient, calculate using the function k

Value

The value returned is the location in millimetres, along the sample, to be cut to obtain a segment representing the number of days indicated by the Time argument

References

Hall-Aspland, S. A., Rogers, T. L., & Canfield, R. B. (2005). Stable carbon and nitrogen isotope analysis reveals seasonal variation in the diet of leopard seals. Marine Ecology Progress Series, 305, 249-259.

Author

The function was written by Gary Truong with collaboration from Ben Walker and Anna Lewis from the University of New South Wales

Examples

LP1(L = 100, L.asym = 150, Time = 28, k = 0.0126)

#> [1] 72.74